Pulse: validate fft data stream before rendering and updating Fling state
Due to a variety of factors, not all audio players access the Android
audio effects stack. Thus, when one of these players starts a track,
the stream may not pass through the visualizer framework. There is no
API to determine this state. So upon linking to visualizer, we analyze
a series of frames to determine if the fft stream is empty. Only if the
steam is valid do we update the Fling logo state and enable drawing flags.

Some general code cleanup as well.
  -- rename some members for consistency
  -- scrap visualizer flash logic, no-op here
  -- Use "mResetDrawing" instead of "mSizeChanged" to trigger
     canvas and bitmap reset. Reset drawing on link and unlink